Fixed SEGFAULT (proximately) caused by a change in compiler behaviour.
In 2017, a global array declared constant as 'const size_t pippo
[PIPPOLEN];' could still have its elements modified by a global pointer
declared as 'size_t *pippo_ptr = (size_t*)pippo;', when compiling with
gcc. Now it would SEGFAULT as it seems that the original array is now
placed into a read-only memory segment by gcc and clang, on GNU/Linux.